requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
groq==0.9.0
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.7.1
python-dotenv==1.0.1
